Set JsonSerializerSettings on every serialize/deserialize call
I found a solution for the problem. It's possible to set JsonSerializerSettings on every call to .SerializeObject(...) and DeserializeObject(...). As the serialization/deserialization logic is centrally hold in BaseDto.cs it's possible to initialize a private static variable of type JsonSerializerSettings in the static constructor and then pass the settings to every call to .SerializeObject(...) and DeserializeObject(...)

resolves #34